requirement.txt
pymongo
python-dotenv
pymongo[srv]
flask-cors
Flask-PyMongo
dnspython